feat!(mkDockerManifest): rewrite logic and args
BREAKING CHANGE: The arguments and the inner logic of `mkDockerManifest` have been rewritten from scratch. The function now also handles the login to the Docker registries. Instead of providing the name (or names) of an image, you now define them for each registry individually. You can now opt out of all automatically assigned tags. There is a new GitHub integration that fetches most information automatically when run in a GitHub action. Please refer to the documentation for more details about the new arguments.